# refactor(ext/http): use scopeguard defer to handle async drop