@lrowe 2023-09-24
I assume the earlier windows test failure were spurious since it passed when running in my fork here: https://github.com/lrowe/deno/actions/runs/6287030911