@lrowe 2023-09-24
Benchmark results seem to show no change in performance.

This branch. `./deno-b129b8950 run --allow-all multi.ts minimal`
```
% third_party/prebuilt/mac/wrk http://localhost:8080/
Running 10s test @ http://localhost:8080/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    77.17us   22.10us 634.00us   73.23%
    Req/Sec    57.49k     1.58k   59.19k    92.57%
  1155607 requests in 10.10s, 165.31MB read
Requests/sec: 114417.97
Transfer/sec:     16.37MB
```

Main. `./deno-cb9ab9c3a run --allow-all multi.ts minimal`
```
% third_party/prebuilt/mac/wrk http://localhost:8080/
Running 10s test @ http://localhost:8080/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    77.25us   22.15us 676.00us   72.89%
    Req/Sec    57.39k     1.35k   59.16k    90.10%
  1153767 requests in 10.10s, 165.05MB read
Requests/sec: 114235.08
Transfer/sec:     16.34MB
```