@lrowe 2023-09-24
Given that the windows tests seem to consistently fail the same tests on this branch when using the xl windows runner I'm going to try to get GitHub actions to run just the failing tests to see and add logging to see what's going on.

```
[serve_test 020.85] httpServerTcpCancellation_file ... FAILED (107ms)
[serve_test 020.90] httpServerTcpCancellation_stream ... FAILED (44ms)
```